集成  redisson

+++ src/main/java/com/springboot/md/redisson/DistributedLocker.java
@@ -7,7 +7,7 @@ import java.util.concurrent.TimeUnit;
 /**
  * @Author: 秒度
  * @Email: fangxin.md@Gmail.com
- * @Date: 2020-12-06 22:54
+ * @Date: 2020-12-11 0:34
  * @Description:
  */
 
@@ -15,13 +15,14 @@ public interface DistributedLocker {
 
     RLock lock(String lockKey);
 
-    RLock lock(String lockKey, int timeout);
+    RLock lock(String lockKey, long timeout);
 
-    RLock lock(String lockKey, TimeUnit unit, int timeout);
+    RLock lock(String lockKey, TimeUnit unit, long timeout);
 
-    boolean tryLock(String lockKey, TimeUnit unit, int waitTime, int leaseTime);
+    boolean tryLock(String lockKey, TimeUnit unit, long waitTime, long leaseTime);
 
     void unlock(String lockKey);
 
     void unlock(RLock lock);
+
 }
